Bug fixes
Fixed a bug that occurred when the api key responsible for refreshing the ranks was deleted. Now removes the server with a failty api key.
Fixed a bug that occurred when an api key with too few permissions was given in installation and then a proper key was given if the user continued the installation.